pyproject.toml
[project]
name = "elevator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
